Re-added glass break sensors
USERS UPGRADING
If you have NOT updated to "v1.6.8", please DO NOT update to this version until you do. Attempts to skip the aforementioned version, WILL crash Homebridge.

To upgrade to `v1.6.8` manually, please enter this command into the Terminal:

`npm install homebridge-adt-pulse@1.6.8`

Then proceed to restart Homebridge for the changes to take effect. Afterwards, you are then safe to upgrade to the latest version.

ADDED
- [Plugin code] Glass break sensors (as occupancy sensors).

PRIVACY NOTICE
Because of unknown bugs plagued in the plugin (that people aren't reporting on GitHub), this version will contain a bug reporting tool that will send debug logs here:

"https://console.re/vHeMZvGjaBWjN29b8PARrCpBNtHTWYteMGFejmxDwgUpan9SMU"

Please be advised these logs are not saved, are ANONYMOUS and in no way, you will be geographically or personally identified. The information that is being sent is:
- A random session number (to group actions).
- Date and time.
- Basic system information (system and architecture version).
- Node version, Homebridge version, Plugin version.
- System status changes (e.g. "all quiet" to "motion").
- Set device logs (e.g. "disarmed" to "away").
- Wrong web portal version messages (e.g. "17.0.0" to "18.0.0").